Refactor: merg ImgPreview, FilePreview
이미지 미리보기 컴포넌트를 파일 미리보기 컴포넌트와 병합
