--- frontend/src/container/ImgPreview/index.js
@@ -1,1 +1,0 @@
-export { default } from './ImgPreview'
